- added 100ms sleep between start and stop - compute average over measured data - altered output data type from long long to float

+++ client/plugins/rapl_plugin.h
@@ -10,7 +10,7 @@ typedef struct RAPL_Plugin_t RAPL_Plugin;
 struct RAPL_Plugin_t
 {
     char events[MAX_RAPL_EVENTS][PAPI_MAX_STR_LEN];
-    long long *values;
+    float *values;
     int data_types[MAX_RAPL_EVENTS];
     int num_events;
 };
